Finish plugin tests and refactor plugin module
