+++ src/conan_app_launcher/ui/views/conan_search/conan_search.py
@@ -2,7 +2,7 @@ from typing import TYPE_CHECKING, Optional
 
 import conan_app_launcher.app as app
 from conan_app_launcher.ui.common.syntax_highlighting import ConfigHighlighter  # using global module pattern
-from conan_app_launcher.ui.plugin.plugins import PluginDescription, PluginInterfaceV1
+from conan_app_launcher.ui.plugin import PluginDescription, PluginInterfaceV1
 from conan_app_launcher.ui.views import LocalConanPackageExplorer
 from conan_app_launcher.ui.widgets import RoundedMenu
 from PySide6.QtCore import QPoint, Qt, Slot, QPropertyAnimation, QEasingCurve
