+++ src/conan_app_launcher/ui/views/package_explorer/package_explorer.py
@@ -2,7 +2,7 @@ import os
 from typing import TYPE_CHECKING
 
 from conan_app_launcher.ui.common.model import re_register_signal
-from conan_app_launcher.ui.plugin.plugins import PluginDescription, PluginInterfaceV1
+from conan_app_launcher.ui.plugin import PluginDescription, PluginInterfaceV1
 from conan_app_launcher.ui.views.package_explorer.controller import PackageFileExplorerController, PackageSelectionController
 from conan_app_launcher.ui.widgets import RoundedMenu
 from PySide6.QtCore import Qt, Signal
