+++ src/conan_app_launcher/ui/views/plugins_manager/plugins.py
@@ -1,7 +1,7 @@
 from typing import TYPE_CHECKING, Optional
 from conan_app_launcher import AUTHOR, BUILT_IN_PLUGIN, DEBUG_LEVEL, __version__
 from PySide6.QtWidgets import QWidget, QFileDialog, QMessageBox
-from conan_app_launcher.ui.plugin.plugins import PluginDescription, PluginHandler, ThemedWidget, PluginInterfaceV1
+from conan_app_launcher.ui.plugin import PluginDescription, PluginHandler, PluginInterfaceV1
 from conan_app_launcher.ui.views.plugins_manager.model import PluginModelItem
 
 from .controller import PluginController
@@ -13,6 +13,7 @@ if TYPE_CHECKING:
 
 
 class PluginsPage(PluginInterfaceV1):
+    """ GUI representation of plugin_handler """
 
     def __init__(self, parent: QWidget, plugin_handler: PluginHandler, base_signals: Optional["BaseSignals"] = None,
                  page_widgets: Optional["FluentWindow.PageStore"] = None):
